Conflicts:
	Despegar.Core.Business/Flight/Itineraries/CheapestPrice.cs
	Despegar.WP.UI/Controls/Flights/FlightSegment.xaml.cs

+++ Despegar.Core.Business/Flight/Itineraries/CheapestPrice.cs
@@ -7,8 +7,8 @@ namespace Despegar.Core.Business.Flight.Itineraries
 {
     public class CheapestPrice
     {
-        public double savings { get; set; }
-        public double total_price { get; set; }
+        public decimal savings { get; set; }
+        public decimal total_price { get; set; }
         public string departure_date { get; set; }
         public string return_date { get; set; }
     }
